Create class to manage consensus collection
