+++ pyrates/consensus.py
@@ -1,10 +1,7 @@
 """Functions to facilitate consensus calling.
 """
 
-import logging
 from collections import defaultdict
-
-import pyrates.sequence as pseq
 import pyrates.utils as utils
 
 class Consensus(object):
@@ -202,127 +199,3 @@ class Consensus(object):
     def __repr__(self):
         return "Consensus(uid=%r, sequence=%r, diffs=%r, size=%r)" % \
                          (self.uid, self.sequence, dict(self.diffs), self.size)
-
-def from_fastq(input_file, id_length, adapter):
-    """Read FASTQ file to generate consensus sequences.
-
-    Args:
-        input_file (:obj:`str`): Name of input file.
-        id_length (:obj:`int`): Length of UID sequence at beginning/end of read.
-        adapter (:obj:`str`): Adapter sequence.
-
-    Returns:
-        :obj:`dict`: Computed consensus sequences.
-    """
-    logger = utils.get_logger(__name__)
-    line_count = 0
-    total_skipped = 0
-    seq = {}
-    id_length = id_length
-    adapt_length = id_length + len(adapter)
-
-    open_fun = utils.smart_open(input_file)
-    with open_fun(input_file) as fastq:
-        for line in fastq:
-            # print out some stats as we go
-            if logger.isEnabledFor(logging.DEBUG) and (line_count % 100000) == 0:
-                logger.debug("reads: %d clusters: %d skipped: %d",
-                             line_count/4, len(seq), total_skipped)
-            elif (line_count % 4) == 1:
-                line = line.rstrip("\n")
-                nameid = line[0:id_length] + line[-id_length:]
-                sequence = line[adapt_length:-adapt_length]
-            elif (line_count % 4) == 3:
-                line = line.rstrip("\n")
-                qnameid = line[0:id_length] + line[-id_length:]
-                qsequence = line[adapt_length:-adapt_length]
-
-                uid = pseq.SequenceWithQuality(nameid, qnameid)
-                read_seq = pseq.SequenceWithQuality(sequence, qsequence)
-                if nameid not in seq:
-                    seq[nameid] = Consensus(uid, read_seq)
-                else:
-                    success = seq[nameid].update(uid, read_seq)
-                    if not success:
-                        total_skipped += 1
-            line_count += 1
-    return seq
-
-def to_fastq(seq, output_file, id_tolerance, merge_size, merge_target):
-    """Write consensus sequences to fastq file, optionally merging clusters.
-
-    If merging of small clusters can be enabled by setting all relevant arguments
-    (`id_tolerance`, `merge_size`, `merge_target`) to positive values. If this is
-    the case clusters smaller than `merge_size` will be compared to all clusters
-    with no more than `merge_target` members. If their UIDs don't differ at more
-    than `id_tolerance` positions an attempt will be made to merge the two clusters,
-    but the attempt may fail if the sequences are too different (using the same
-    logic employed when computing the initial consensus).
-
-    Args:
-        seq (:obj:`dict`): Read clusters and their consensus sequences, identified
-            by the sequence of their UIDs.
-        output_file (:obj:`str`): File name for output. Will be replaced if it exists.
-        id_tolerance (:obj:`int`): Maximum number of mismatches between UIDs allowed
-            for merging of clusters.
-        merge_size (:obj:`int`): Maximum size for *small* clusters that will be considered
-            for merging.
-        merge_target (:obj:`int`): Only clusters that are not larger than this are
-            considered when trying to find matches for small clusters.
-    """
-    logger = utils.get_logger(__name__)
-    output_fun = utils.smart_open(output_file)
-    with output_fun(output_file, 'w') as output:
-        if merge_size and merge_target:
-            logger.info('Merging small clusters')
-            candidates = []
-            targets = []
-            merge_count = 0
-            for (seq_count, uid) in enumerate(seq):
-                if logger.isEnabledFor(logging.DEBUG) and seq_count % 10000 == 0:
-                    logger.debug("clusters: %d, merged: %d, small: %d, targets: %d",
-                                 seq_count, merge_count, len(candidates), len(targets))
-                if seq[uid].size <= merge_size:
-                    merged = False
-                    for consensus in targets:
-                        merged = consensus.merge(seq[uid], id_tolerance)
-                        if merged:
-                            merge_count += 1
-                            break
-                    if not merged:
-                        ## attempt merging with other candidates
-                        remove = None
-                        for (i, cand) in enumerate(candidates):
-                            merged = seq[uid].merge(cand, id_tolerance)
-                            if merged:
-                                merge_count += 1
-                                remove = i
-                                break
-                        if merged:
-                            del candidates[remove]
-                            if seq[uid].size > merge_size:
-                                targets.append(seq[uid])
-                            else:
-                                candidates.append(seq[uid])
-                        else:
-                            candidates.append(seq[uid])
-                elif seq[uid].size <= merge_target:
-                    targets.append(seq[uid])
-                    processed = []
-                    for (i, cand) in enumerate(candidates):
-                        if seq[uid].merge(cand, id_tolerance):
-                            processed.insert(0, i)
-                    for i in processed:
-                        del candidates[i]
-                        merge_count += 1
-                else:
-                    output.write(str(seq[uid]) + "\n")
-            logger.info('Clusters merged: %d', merge_count)
-            logger.info('Small clusters remaining: %d', len(candidates))
-            for consensus in targets:
-                output.write(str(consensus) + "\n")
-            for consensus in candidates:
-                output.write(str(consensus) + "\n")
-        else:
-            for uid in seq:
-                output.write(str(seq[uid]) + "\n")
